+++ tests/device/screen2_data/ioc.yaml
@@ -1,0 +1,19 @@
+---
+YAGS:IN20:351:Acquire:
+  value: 0
+  dispatch:
+    YAGS:IN20:351:IMAGE:
+      0: [0, 0, 0, 0]
+      1: [0, 1, 2, 3]
+YAGS:IN20:351:IMAGE: [0, 0, 0, 0]
+YAGS:IN20:351:N_OF_ROW: 2
+YAGS:IN20:351:N_OF_COL: 2
+YAGS:IN20:351:PNEUMATIC:
+  dispatch:
+    YAGS:IN20:351:TGT_STS:
+      0: 1
+      1: 2
+YAGS:IN20:351:TGT_STS: 1
+YAGS:IN20:211:TGT_STS: 1
+# Upstream device is in
+YAGS:IN20:241:TGT_STS: 2
